+types, to slim down extras and have basic set of types with minimal dependencies
This also allows to remove some import loops to do further cleanups,
like moving save and load functions out of extras.
