+++ accelerator/test_methods/a_test_hashlabel.py
@@ -24,7 +24,7 @@ wrong hashlabel without rehashing is not allowed.
 '''
 
 from accelerator.dataset import DatasetWriter
-from accelerator.extras import DotDict
+from accelerator import DotDict
 from accelerator.dsutil import typed_writer
 from accelerator.error import DatasetUsageError
 from accelerator.compat import unicode
